// frontend/src/api/generated/models/syncRunStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type SyncRunStatus = typeof SyncRunStatus[keyof typeof SyncRunStatus];


export const SyncRunStatus = {
  Pending: 'Pending',
  Running: 'Running',
  Succeeded: 'Succeeded',
  Failed: 'Failed',
} as const;